# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for AI Recipes, an Indian-first library of practical AI workflows.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: Search and filter published recipes
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: audience, in: query, schema: { type: string } }
        - { name: difficulty, in: query, schema: { type: string } }
        - { name: tool, in: query, schema: { type: string } }
        - { name: language, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [relevance, popular, newest] } }
        - { name: page, in: query, schema: { type: number, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: number, minimum: 1, maximum: 50, default: 12 } }
      responses:
        "200":
          description: Recipe results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeListResponse"
  /recipes/{slug}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe by slug
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "404":
          description: Recipe not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /recipes/{slug}/related:
    get:
      operationId: getRelatedRecipes
      tags: [recipes]
      summary: Get related published recipes
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: number, minimum: 1, maximum: 6, default: 3 } }
      responses:
        "200":
          description: Related recipes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RecipeCard" }
...
  /categories/{slug}:
    get:
      operationId: getCategory
      tags: [discovery]
      summary: Get category details and recipes
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionResponse" }
...
  /audiences/{slug}:
    get:
      operationId: getAudience
      tags: [discovery]
      summary: Get audience details and recipes
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Audience
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionResponse" }
  /events:
    post:
      operationId: trackEvent
      tags: [engagement]
      summary: Track a lightweight product event
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventInput" }
      responses:
        "204": { description: Event accepted }
  /feedback:
    post:
      operationId: submitFeedback
      tags: [engagement]
      summary: Submit recipe feedback
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackInput" }
      responses:
        "201":
          description: Feedback saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /suggestions:
    post:
      operationId: submitSuggestion
      tags: [submissions]
      summary: Suggest a recipe for a missing search need
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SuggestionInput" }
      responses:
        "201":
          description: Suggestion saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /submissions:
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit an AI workflow for review
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmissionInput" }
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /me:
    get:
      operationId: getMe
      tags: [engagement]
      summary: Get signed-in profile summary
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401": { description: Authentication required }
  /me/saved:
    get:
      operationId: listSavedRecipes
      tags: [engagement]
      summary: List the signed-in user's saved recipes
      responses:
        "200":
          description: Saved recipes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RecipeCard" }
        "401": { description: Authentication required }
  /me/saved/{recipeId}:
    post:
      operationId: saveRecipe
      tags: [engagement]
      summary: Save a recipe
      parameters:
        - { name: recipeId, in: path, required: true, schema: { type: number } }
      responses:
        "201": { description: Recipe saved }
        "401": { description: Authentication required }
    delete:
      operationId: unsaveRecipe
      tags: [engagement]
      summary: Remove a saved recipe
      parameters:
        - { name: recipeId, in: path, required: true, schema: { type: number } }
      responses:
        "204": { description: Recipe unsaved }
        "401": { description: Authentication required }
  /me/history:
    get:
      operationId: listHistory
      tags: [engagement]
      summary: List recently viewed recipes
      responses:
        "200":
          description: Recently viewed recipes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RecipeCard" }
        "401": { description: Authentication required }
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get dashboard metrics
      responses:
        "200":
          description: Admin metrics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSummary" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/recipes:
    get:
      operationId: listAdminRecipes
      tags: [admin]
      summary: List recipes for admin
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: number, minimum: 1, default: 1 } }
      responses:
        "200":
          description: Admin recipe list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeListResponse"
    post:
      operationId: createAdminRecipe
      tags: [admin]
      summary: Create an admin recipe
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecipeInput" }
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
  /admin/recipes/{id}:
    patch:
      operationId: updateAdminRecipe
      tags: [admin]
      summary: Update an admin recipe
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecipeUpdate" }
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
  /admin/recipes/{id}/duplicate:
    post:
      operationId: duplicateAdminRecipe
      tags: [admin]
      summary: Duplicate a recipe as a draft
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "201":
          description: Duplicated recipe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    MessageResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    RecipeCard:
      type: object
      properties:
        id: { type: number }
        title: { type: string }
        slug: { type: string }
        shortDescription: { type: string }
        category: { type: string }
        categorySlug: { type: string }
        audience: { type: string }
        audienceSlug: { type: string }
        difficulty: { type: string }
        estimatedTime: { type: string }
        language: { type: string }
        tags:
          type: array
          items: { type: string }
        featured: { type: boolean }
        trending: { type: boolean }
        viewCount: { type: number }
        copyCount: { type: number }
        saveCount: { type: number }
      required: [id, title, slug, shortDescription, category, categorySlug, audience, audienceSlug, difficulty, estimatedTime, language, tags, featured, trending, viewCount, copyCount, saveCount]
    Recipe:
      allOf:
        - $ref: "#/components/schemas/RecipeCard"
        - type: object
          properties:
            problem: { type: string }
            subcategory: { type: string }
            aiTools:
              type: array
              items: { type: string }
            requiredInputs:
              type: array
              items: { type: string }
            steps:
              type: array
              items: { type: string }
            prompt: { type: string }
            exampleInput: { type: string }
            exampleOutput: { type: string }
            refinementPrompts:
              type: array
              items: { type: string }
            verificationNotes: { type: string }
            status: { type: string }
            version: { type: string }
            testedAt: { type: ["string", "null"], format: date }
            testedWith:
              type: array
              items: { type: string }
            seoTitle: { type: string }
            seoDescription: { type: string }
            publishedAt: { type: ["string", "null"], format: date-time }
          required: [problem, subcategory, aiTools, requiredInputs, steps, prompt, exampleInput, exampleOutput, refinementPrompts, verificationNotes, status, version, testedWith, seoTitle, seoDescription]
...
    RecipeListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/RecipeCard" } }
        page: { type: number }
        pageSize: { type: number }
        total: { type: number }
        query: { type: string }
      required: [items, page, pageSize, total, query]
    CollectionResponse:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        recipes: { type: array, items: { $ref: "#/components/schemas/RecipeCard" } }
      required: [name, slug, description, recipes]
...
    EventInput:
      type: object
      properties:
        type: { type: string, enum: [page_view, recipe_view, recipe_copy, recipe_save, recipe_unsave, recipe_share, recipe_feedback_positive, recipe_feedback_negative, search, recipe_submission] }
        recipeId: { type: ["integer", "null"] }
        query: { type: ["string", "null"], maxLength: 200 }
      required: [type]
...